+++ src/components/task-new-button.vue
@@ -1,13 +1,13 @@
 <template>
   <div>
     <q-modal v-model="dialogOpened"
-             content-css="min-width:60vw;min-height:300px"
+             content-css="min-width:60vw;min-height:500px"
              no-backdrop-dismiss
     >
       <q-modal-layout>
         <q-toolbar slot="header">
           <q-toolbar-title>
-            Add parameter
+            Register new task
           </q-toolbar-title>
           <q-btn round dense
                  color="primary"
@@ -18,13 +18,13 @@
 
         <div class="q-pa-lg">
           <q-field
-            label="Name:"
-            :error="!name.length"
+            label="Executor ID:"
+            :error="!executorId.length"
             :label-width="4"
           >
             <q-input
               class="bg-light-blue-1"
-              v-model.trim="name"
+              v-model.trim="executorId"
               type="text"
               autofocus
               style="width:24em"
@@ -32,48 +32,50 @@
           </q-field>
 
           <q-field
-            label="Type:"
-            :error="!type.length"
+            label="Task Definition ID:"
+            :error="!taskDefId.length"
             :label-width="4"
           >
             <q-input
               class="bg-light-blue-1"
-              v-model.trim="type"
+              v-model.trim="taskDefId"
               type="text"
               style="width:24em"
             />
           </q-field>
 
           <q-field
-            label="Description:"
+            label="Asset ID:"
+            :error="!assetId.length"
+            :label-width="4"
+          >
+            <q-input
+              class="bg-light-blue-1"
+              v-model.trim="assetId"
+              type="text"
+              style="width:24em"
+            />
+          </q-field>
+
+          <q-field
+            label="Task Name:"
+            :label-width="4"
+          >
+            <q-input
+              class="bg-light-blue-1"
+              v-model.trim="name"
+              type="text"
+              style="width:24em"
+            />
+          </q-field>
+
+          <q-field
+            label="Task Description:"
             :label-width="4"
           >
             <q-input
               class="bg-light-blue-1"
               v-model.trim="description"
-              type="text"
-              style="width:24em"
-            />
-          </q-field>
-
-          <q-field
-            label="Required?"
-            :label-width="4"
-            :warning="required"
-          >
-            <q-checkbox
-              v-model="required"
-              :left-label="true"
-            />
-          </q-field>
-
-          <q-field
-            label="Default Value:"
-            :label-width="4"
-          >
-            <q-input
-              class="bg-light-blue-1"
-              v-model.trim="defaultValue"
               type="text"
               style="width:24em"
             />
@@ -108,11 +110,7 @@ import { Notify } from 'quasar'
 
 export default {
   props: {
-    executorId: {
-      type: String,
-      required: true
-    },
-    taskdefId: {
+    planId: {
       type: String,
       required: true
     },
@@ -125,44 +123,63 @@ export default {
   data () {
     return {
       dialogOpened: false,
+      executorId: '',
+      taskDefId: '',
+      assetId: '',
       name: '',
-      type: '',
-      required: true,
-      defaultValue: '',
-      description: ''
+      description: '',
+      arguments: [],
+      startDate: null,
+      endDate: null
     }
   },
 
   computed: {
     okToSubmit () {
-      return this.name && this.type
+      return this.taskDefId && this.assetId
     }
   },
 
   methods: {
     openDialog () {
+      this.taskDefId = ''
+      this.assetId = ''
       this.name = ''
-      this.type = ''
-      this.required = true
-      this.defaultValue = ''
       this.description = ''
+      this.arguments = []
+      this.startDate = new Date()
+      this.endDate = new Date()
       this.dialogOpened = true
     },
 
     submit () {
       const data = {
-        name: this.name,
-        type: this.type,
-        required: this.required
+        executorId: this.executorId,
+        taskDefId: this.taskDefId,
+        assetId: this.assetId
       }
-      if (this.defaultValue) {
-        data.defaultValue = this.defaultValue
+      if (this.name) {
+        data.name = this.name
       }
       if (this.description) {
         data.description = this.description
       }
+      // TODO arguments
+      // if (this.arguments) {
+      //   data.arguments = ...
+      // }
+      if (this.startDate) {
+        data.start = this.startDate.toISOString()
+      }
+      if (this.endDate) {
+        data.end = this.endDate.toISOString()
+      }
+      // TODO geometry
+      // if (this.geometry) {
+      //   data.geometry = ...
+      // }
 
-      const url = `/executors/${this.executorId}/taskdefs/${this.taskdefId}/parameters`
+      const url = `/plans/${this.planId}/tasks`
       this.$axios({
         method: 'POST',
         url,
@@ -172,7 +189,7 @@ export default {
           console.log(`POST ${url}: response=`, response)
           this.dialogOpened = false
           Notify.create({
-            message: 'Parameter created',
+            message: 'Plan task registered',
             timeout: 1000,
             type: 'info'
           })