+++ src/pages/Plan.vue
@@ -13,7 +13,7 @@
     <div v-if="plan">
       <h5> Plan: {{plan.name}}
       </h5>
-      <small>{{params.planId}}</small>
+      <small>{{plan.planId}}</small>
 
       <div>
         Description: {{ plan.description}}
@@ -22,10 +22,15 @@
       <div>
         Tasks:
         <ul>
+          <li>
+            <task-new-button
+              :plan-id="plan.planId"
+              :created="created"
+            />
+          </li>
           <li v-for="task in plan.tasks" :key="task.taskId">
-            <!-- Note: showing the associated taskDefId -->
             <router-link :to="`/plans/${plan.planId}/tasks/${task.taskId}`">
-              {{task.taskDefId}}</router-link>
+              {{task.name || task.taskDefId || task.taskId}}</router-link>
 
             <span v-if="task.arguments && task.arguments.length">
               ( {{task.arguments.length }} explicit arguments)
@@ -43,10 +48,15 @@
 </template>
 
 <script>
+import TaskNewButton from 'components/task-new-button'
 import lodash from 'lodash'
 const _ = lodash
 
 export default {
+  components: {
+    TaskNewButton
+  },
+
   data () {
     return {
       loading: false,
@@ -78,11 +88,18 @@ export default {
           this.loading = false
           console.log(`GET ${url}: response=`, response)
           this.plan = response.data || {}
+          if (!this.plan.tasks) {
+            this.plan.tasks = []
+          }
         })
         .catch(e => {
           this.loading = false
           console.error(e)
         })
+    },
+
+    created (data) {
+      this.plan.tasks.splice(0, 0, data)
     }
   },