+++ src/pages/TaskDef.vue
@@ -3,8 +3,9 @@
     <q-breadcrumbs active-color="secondary" color="light">
       <q-breadcrumbs-el label="Home" to="/" />
       <q-breadcrumbs-el label="Executors" to="/executors" />
-      <q-breadcrumbs-el label="Executor" :to="`/executors/${params.executorId}`" />
-      <q-breadcrumbs-el label="TaskDef" />
+      <q-breadcrumbs-el :label="params.executorId" :to="`/executors/${params.executorId}`" />
+      <q-breadcrumbs-el label="TaskDefs" />
+      <q-breadcrumbs-el :label="params.taskDefId" />
       <q-btn
         dense round icon="refresh" class="q-ml-lg" size="sm"
         @click="refresh"
@@ -12,14 +13,18 @@
     </q-breadcrumbs>
 
     <div v-if="taskDef">
-      <h5> Task Definition: {{params.taskDefId}}
+
+      <h5> Task Definition: {{ params.taskDefId }}
       </h5>
+
       <div>
-        Description: {{ taskDef.description}}
+        Description: {{ taskDef.description }}
       </div>
+
       <div>
-        Asset classes: {{ join(taskDef.assetClasses)}}
+        Asset classes: {{ join(taskDef.assetClasses) }}
       </div>
+
       <q-table
         title="Parameters"
         :data="tableData"